Move setting prescribed CH4 to just after emissions
It is necessary to move where prescribed CH4 is set because it is
dependent on PBL height and previously PBL height was not updated until
after Set_CH4 was called. This introduces small differences in GC-Classic
when breaking up runs. CH4 is now set using updated PBL height and is
grouped with emissions where other prescribed surface VMRs are applied.
It is still applied outside of emissions.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>