No diff fixes to SOIL_WETDEP and SOIL_DRYDEP
This commit removes the level argument since it is not actually used in
the subroutines. This update also uncomments a Soil NOx extension logical
bracketed around a call to SOIL_WETDEP. The subroutine only changes
arrays used in Soil NOx extension so calling it when the extension is off
is not necessary.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>